Fraud Detection/requirements.txt
ExifRead==2.3.1
Pillow==8.0.0
matplotlib==3.3.2
numpy==1.19.2
